User Permission is updated and pushed to staging.

+++ emergency_slip_record.php
@@ -58,22 +58,26 @@
                               <i class='fas fa-wallet'></i> Bill</a>
                               <br> 
                               <a href='emergency_slip_print.php?sid=$rs[SLIP_ID]' style='color:green;'>
-                              <i class='fas fa-wallet'></i> Print</a>
-                              <br>
+                              <i class='fas fa-wallet'></i> Print</a>";
+                              if ($_SESSION['type'] == "admin") {  
+                              echo "<br>
                               <a href='add_patient.php?id=$rs[SLIP_ID]'><i class='fas fa-edit'></i> Edit</a>
                               <br>
                               <a onClick=\"javascript: return confirm('Please confirm deletion');\" 
                               href='backend_components/delete_handler.php?esrId=$rs[SLIP_ID]' style='color:red;'>
                               <i class='fas fa-trash'></i> Delete</a>";
+                              }
                         }else{
                             echo "<a href='emergency_slip_print.php?sid=$rs[SLIP_ID]' style='color:green;'>
-                            <i class='fas fa-wallet'></i> Print</a>
-                            <br>
+                            <i class='fas fa-wallet'></i> Print</a>";
+                            if ($_SESSION['type'] == "admin") {  
+                            echo "<br>
                             <a href='add_patient.php?id=$rs[SLIP_ID]'><i class='fas fa-edit'></i> Edit</a>
                             <br>
                             <a onClick=\"javascript: return confirm('Please confirm deletion');\" 
                             href='backend_components/delete_handler.php?esrId=$rs[SLIP_ID]' style='color:red;'>
                             <i class='fas fa-trash'></i> Delete</a>";
+                            }
                         }
                           
                         echo "</td></tr>";
